# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Avito.Kitchen Platform API
  description: >
    API платформы Авито.Кухня: каталог заведений, меню, корзина, заказы.

    Аутентификация не реализована (по условиям задания) — пользователь
    идентифицируется заголовком X-User-Id. При первом обращении с новым
    X-User-Id пользователь автоматически заводится в БД (upsert),
    отдельного signup не требуется.

    Заведения считаются предзаполненными (сидинг через миграции),
    self-registration не предусмотрен. Адрес для доставки вебхуков
    заведению (webhook_url) хранится в доменной модели заведения и
    в БД, но не управляется через публичное API в рамках MVP.

    Межсервисная авторизация (проверка, что PATCH /orders/{id}/status
    вызывает именно "тот самый" сервис заведения, а не любой желающий)
    не реализована — это осознанное упрощение MVP, следуя условиям
    задания (авторизацию не делать).
  version: 0.2.0
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /restaurants/{restaurantId}/menu:
    get:
      operationId: getRestaurantMenu
      tags: [Menu]
      summary: Получить меню заведения (категории с блюдами)
      description: >
        Возвращает только "живые" блюда (deletedAt IS NULL). Блюда,
        временно отсутствующие (available=false), остаются в выдаче,
        чтобы клиент мог показать их как недоступные, а не скрывать.
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      responses:
        '200':
          description: Меню заведения
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryWithDishes'
        '404':
          $ref: '#/components/responses/NotFound'
  /restaurants/{restaurantId}/categories:
    post:
      operationId: createCategory
      tags: [Menu]
      summary: Добавить категорию меню
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                sortOrder:
                  type: integer
      responses:
        '201':
          description: Категория создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /restaurants/{restaurantId}/categories/{categoryId}:
    delete:
      operationId: deleteCategory
      tags: [Menu]
      summary: Удалить категорию
      description: >
        Запрещено, если в категории есть хотя бы одно блюдо — включая
        мягко удалённые (deletedAt IS NOT NULL), т.к. такое блюдо можно
        восстановить и оно должно на что-то ссылаться. Сначала нужно
        перенести (PATCH categoryId) или полностью не иметь блюд в
        категории.
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Категория удалена
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: В категории есть блюда (в т.ч. удалённые), удаление запрещено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurantId}/dishes:
    post:
      operationId: createDish
      tags: [Menu]
      summary: Добавить блюдо
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishInput'
      responses:
        '201':
          description: Блюдо создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '404':
          description: categoryId не найден среди категорий этого заведения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurantId}/dishes/{dishId}:
    patch:
      operationId: updateDish
      tags: [Menu]
      summary: Изменить название/описание/цену/категорию блюда
      description: >
        Позволяет в т.ч. перенести блюдо в другую категорию (categoryId).
        Если указанная categoryId существует, но принадлежит другому
        restaurantId — возвращается 404 (категория считается
        "не найденной" в контексте этого заведения, а не 409).
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/DishId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId:
                  type: string
                  format: uuid
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                  format: float
                  minimum: 0
      responses:
        '200':
          description: Блюдо обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '404':
          description: Блюдо не найдено, либо categoryId не принадлежит этому заведению
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteDish
      tags: [Menu]
      summary: Удалить блюдо (soft-delete)
      description: >
        Проставляет deletedAt, физически строка не удаляется (нужна
        для ссылок из старых заказов и восстановления). После удаления
        блюдо пропадает из GET /menu и недоступно для добавления в
        корзину (трактуется как unavailable при checkout существующих
        позиций в чужих корзинах).
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/DishId'
      responses:
        '204':
          description: Блюдо удалено
        '404':
          $ref: '#/components/responses/NotFound'
  /restaurants/{restaurantId}/dishes/{dishId}/restore:
    post:
      operationId: restoreDish
      tags: [Menu]
      summary: Восстановить ранее удалённое блюдо
      description: >
        Снимает deletedAt. Поле available при этом не меняется —
        остаётся таким, каким было на момент удаления. Если блюдо
        не было удалено, возвращается 404.
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/DishId'
      responses:
        '200':
          description: Блюдо восстановлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '404':
          description: Блюдо не найдено или не было удалено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurantId}/dishes/{dishId}/availability:
    patch:
      operationId: setDishAvailability
      tags: [Menu]
      summary: Изменить наличие блюда (в наличии / закончилось)
      description: Применимо только к неудалённым блюдам (deletedAt IS NULL).
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/DishId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [available]
              properties:
                available:
                  type: boolean
      responses:
        '200':
          description: Наличие обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    RestaurantId:
      name: restaurantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DishId:
      name: dishId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        restaurantId:
          type: string
          format: uuid
        name:
          type: string
        sortOrder:
          type: integer
    Dish:
      type: object
      properties:
        id:
          type: string
          format: uuid
        restaurantId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        available:
          type: boolean
        deletedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          description: null, если блюдо не удалено
    DishInput:
      type: object
      required: [categoryId, name, price]
      properties:
        categoryId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
          minimum: 0
    CategoryWithDishes:
      allOf:
        - $ref: '#/components/schemas/Category'
        - type: object
          properties:
            dishes:
              type: array
              items:
                $ref: '#/components/schemas/Dish'
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: >
            Машиночитаемый код ошибки, напр. CART_MULTI_RESTAURANT,
            RESTAURANT_CLOSED, CATEGORY_NOT_EMPTY, INVALID_TRANSITION
        message:
          type: string
...